requirements.txt
fastapi
uvicorn
python-multipart
streamlit
moviepy
faster-whisper
pillow
psutil
